# Conflicts:
#	HITforISProject-master/IS_Proj_HIT/Views/PCA/CreateAssessment.cshtml
#	HITforISProject-master/IS_Proj_HIT/wwwroot/css/site.css

+++ HITforISProject-master/IS_Proj_HIT/Pages/PCA/PainParameter/Delete.cshtml.cs
@@ -8,6 +8,9 @@ using Microsoft.EntityFrameworkCore;
 using IS_Proj_HIT.Models.Data;
 using IS_Proj_HIT.Models.PCA;
 using Microsoft.AspNetCore.Authorization;
+using System.Transactions;
+using System.Data.Common;
+
 
 namespace IS_Proj_HIT
 {
@@ -27,13 +30,13 @@ namespace IS_Proj_HIT
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            ViewData["RegularMessage"] = "Are you sure you want to delete this?";
+            ViewData["ErrorMessage"] = "";
 
             PainParameter = await _context.PainParameter
-                .Include(p => p.PainScaleType).FirstOrDefaultAsync(m => m.PainParameterId == id);
+                .Include(p => p.PainScaleType)
+                .Include(pr=>pr.PainRatings)
+                .FirstOrDefaultAsync(m => m.PainParameterId == id);
 
             if (PainParameter == null)
             {
@@ -49,14 +52,36 @@ namespace IS_Proj_HIT
                 return NotFound();
             }
 
-            PainParameter = await _context.PainParameter.FindAsync(id);
+            PainParameter = await _context.PainParameter
+                            .Include(p => p.PainScaleType)
+                            .Include(pr => pr.PainRatings)
+                            .FirstOrDefaultAsync(m => m.PainParameterId == id);
 
             if (PainParameter != null)
             {
-                _context.PainParameter.Remove(PainParameter);
-                await _context.SaveChangesAsync();
-            }
+                // See if there are any pain assessments using this pain parameter
+                bool paExists = _context.PcaPainAssessment.Any(p => p.PainParameterId == PainParameter.PainScaleTypeId);
+                if (paExists)
+                {
+                    Console.WriteLine("Pain assessment records exist using these records.");
+                    ViewData["RegularMessage"] = "";
+                    ViewData["ErrorMessage"] = "Pain assessment records exist using these records. Delete not available.";
+                    return Page();
+                }
 
+                using (var tran = new TransactionScope())
+                {
+                    foreach (PainRating pr in PainParameter.PainRatings)
+                    {
+                        _context.PainRating.Remove(pr);
+                    }
+
+                    _context.PainParameter.Remove(PainParameter);
+                    _context.SaveChanges();
+
+                    tran.Complete();
+                }
+            }
             return RedirectToPage("./Index");
         }
     }
